+++ transformers/column.go
@@ -1,0 +1,53 @@
+package transformers
+
+import (
+	"fmt"
+	"reflect"
+
+	"github.com/cloudquery/plugin-sdk/v3/schema"
+)
+
+func (t *structTransformer) addColumnFromField(field reflect.StructField, parent *reflect.StructField) error {
+	if t.ignoreField(field) {
+		return nil
+	}
+
+	columnType, err := t.getColumnType(field)
+	if err != nil {
+		return fmt.Errorf("failed to transform type for field %s: %w", field.Name, err)
+	}
+
+	name, path, err := t.getFieldNamePath(field, parent)
+	if err != nil || name == "" {
+		return err
+	}
+	if t.table.Columns.Get(name) != nil {
+		return nil
+	}
+
+	resolver := t.resolverTransformer(field, path)
+	if resolver == nil {
+		resolver = DefaultResolverTransformer(field, path)
+	}
+
+	column := schema.Column{
+		Name:          name,
+		Type:          columnType,
+		Resolver:      resolver,
+		IgnoreInTests: t.ignoreInTestsTransformer(field),
+	}
+
+	for _, pk := range t.pkFields {
+		if pk == path {
+			// use path to allow the following
+			// 1. Don't duplicate the PK fields if the unwrapped struct contains a fields with the same name
+			// 2. Allow specifying the nested unwrapped field as part of the PK.
+			column.PrimaryKey = true
+			t.pkFieldsFound = append(t.pkFieldsFound, pk)
+		}
+	}
+
+	t.table.Columns = append(t.table.Columns, column)
+
+	return nil
+}
